git-svn-id: http://cadia.ru.is/svn/repos/replicode@366 36b273a1-5dce-4f9d-8980-b90fca7d7570

+++ r_exec/pattern_extractor.cpp
@@ -595,7 +595,7 @@ namespace	r_exec{
 		mdl->code(write_index)=Atom::Set(1);	// only one group: the one the tpx lives in.
 		mdl->code(++write_index)=Atom::RPointer(2);
 
-		mdl->code(MDL_STRENGTH)=Atom::Float(1);
+		mdl->code(MDL_STRENGTH)=Atom::Float(0);
 		mdl->code(MDL_CNT)=Atom::Float(1);
 		mdl->code(MDL_SR)=Atom::Float(1);
 		mdl->code(MDL_DSR)=Atom::Float(1);
